Reorganize root certs by attestation format
